Refactor note hook
