+++ src/features/chats/chat-list/components/chat-list-item/components/chat-content/index.tsx
@@ -1,21 +1,22 @@
+import { useState } from 'react';
 import ReactPlayer from 'react-player';
-
-import { loader } from '@/shared/utils';
 
 import { ChatEvent } from '@/features/chats';
 import { UserMention } from '@/features/users/user-mention';
+
+import { loader } from '@/shared/utils';
 
 import { CategorizedChatContent } from '../../types';
 
 export const ChatContent = ({
   categorizedChatContent,
   sameAsCurrentUser,
-  setSelectedImage,
 }: {
   categorizedChatContent: CategorizedChatContent[];
-  setSelectedImage: React.Dispatch<React.SetStateAction<string | null>>;
   sameAsCurrentUser?: boolean;
 }) => {
+  const [selectedImage, setSelectedImage] = useState<string | null>(null);
+
   return categorizedChatContent.map((part, i) => {
     switch (part.category) {
       case 'text':
@@ -26,13 +27,26 @@ export const ChatContent = ({
         );
       case 'image':
         return (
-          <img
-            key={i}
-            src={loader(part.content, { w: 200 })}
-            alt="chat"
-            className="max-w-full h-40 object-contain rounded-lg mt-2 cursor-pointer"
-            onClick={() => setSelectedImage(part.content)}
-          />
+          <div key={i}>
+            <img
+              src={loader(part.content, { w: 200 })}
+              alt="chat"
+              className="max-w-full mx-auto h-40 object-contain rounded-lg mt-2 cursor-pointer"
+              onClick={() => setSelectedImage(part.content)}
+            />
+            {selectedImage && (
+              <div
+                className="fixed max-w-full p-20 inset-0 z-50 flex items-center justify-center bg-black/80"
+                onClick={() => setSelectedImage(null)}
+              >
+                <img
+                  src={selectedImage}
+                  alt="Enlarged chat"
+                  className="h-auto max-h-svh rounded-lg"
+                />
+              </div>
+            )}
+          </div>
         );
       case 'video':
         return (
